Corregido colores de alumno_Ver2 y dataTable Traducido

+++ src/main/java/pe/edu/sistemas/sismanweb/dao/CategoriaDocenteDAO.java
@@ -14,6 +14,6 @@ public interface CategoriaDocenteDAO {
 	
 	public List<CategoriaDocente> obtenerTodoCategoriaDocente();
 	
-	public CategoriaDocente obtenerCategoriaDocentexID(Integer idCategoriaDocente);	
+	public CategoriaDocente obtenerCategoriaDocentexID(Short idCategoriaDocente);	
 
 }
